integrate with new projects page
