+++ frontend/src/components/ProjectsList.tsx
@@ -1,10 +1,12 @@
 import type { Project } from "../database/models/metadata";
-import { useState, useEffect } from "preact/hooks";
+import { useState, useEffect, useMemo } from "preact/hooks";
 import { Timestamp } from "firebase/firestore";
 import { listProjects } from "../database/firestore";
 import { auth } from "../utils/firebase/firebase-init";
 import { onAuthStateChanged } from "firebase/auth";
 import NewProjectModal from "./NewProjectModal";
+import exitIcon from "../media/cross.png";
+import projectIcon from "../media/project.png";
 
 // Set to true to use hardcoded dummy projects instead of fetching from DB
 const USE_DUMMY_DATA = false;
@@ -45,10 +47,16 @@ const dummyProjects: Project[] = [
     },
 ];
 
-function ProjectsList() {
+interface ProjectsListProps {
+    activeProjectId?: string;
+    onProjectSelect?: (projectId: string) => void;
+}
+
+function ProjectsList({ activeProjectId, onProjectSelect }: ProjectsListProps) {
     const [projects, setProjects] = useState<Project[]>([]);
     const [loading, setLoading] = useState(true);
     const [showNewProjectModal, setShowNewProjectModal] = useState(false);
+    const [query, setQuery] = useState("");
 
     useEffect(() => {
         // Use dummy data for testing
@@ -77,32 +85,79 @@ function ProjectsList() {
         return () => unsubscribe();
     }, []);
 
-    return (
-        <aside 
-            class="bg-[#808080] w-[250px] min-h-[calc(100vh-80px)] py-5 fixed left-0 top-[80px] z-[100] shadow-[2px_0_5px_rgba(0,0,0,0.1)] overflow-y-auto flex flex-col"
-        >
-            <div class="px-5 pb-5 border-b border-white/20 mb-5">
-                <h2 class="text-[#F7F9F1] text-2xl font-semibold m-0">Projects</h2>
-            </div>
-            
-            <ul class="list-none flex-1">
-                {loading ? (
-                    <li class="px-5 py-3 text-[#F7F9F1] opacity-70">Loading...</li>
-                ) : projects.length === 0 ? (
-                    <li class="px-5 py-3 text-[#F7F9F1] opacity-70">No projects yet</li>
-                ) : (
-                    projects.map((project) => (
-                        <li key={project.id} class="flex items-center px-5 py-3 cursor-pointer transition-[background-color] duration-200 ease-in-out text-[#F7F9F1] hover:bg-white/10">
-                            <i class="bx bx-folder text-2xl mr-3" style="color: #EDD16E;"></i>
-                            <span class="text-base font-normal flex-1">{project.name}</span>
-                        </li>
-                    ))
-                )}
-            </ul>
+    // Filter projects based on search query
+    const filtered = useMemo(() => {
+        const q = query.trim().toLowerCase();
+        if (!q) return projects;
+        return projects.filter((p) => p.name.toLowerCase().includes(q));
+    }, [projects, query]);
 
-            <div class="px-5 py-4 border-t border-white/20 mt-auto">
-                <button class="bg-[#037F6F] text-[#F7F9F1] px-4 py-2 rounded-lg text-sm font-bold cursor-pointer border-none w-full hover:bg-[#05B39C] transition-colors duration-300" onClick={() => setShowNewProjectModal(true)}>
-                    + Add Project
+    return (
+        <>
+            {/* Header with title and search */}
+            <div class="projects-sidebar-header">
+                <div class="projects-sidebar-title">Projects</div>
+                <div class="projects-search">
+                    <input
+                        class="projects-search-input"
+                        placeholder="Search Projects"
+                        value={query}
+                        onInput={(e) => setQuery((e.currentTarget as HTMLInputElement).value)}
+                    />
+                    <button
+                        type="button"
+                        class="projects-search-clear"
+                        onClick={() => setQuery("")}
+                        aria-label="Clear search"
+                        title="Clear"
+                    >
+                        <img src={exitIcon} alt="Clear search" class="projects-search-clear-icon" />
+                    </button>
+                </div>
+            </div>
+
+            {/* Projects list */}
+            <div class="projects-sidebar-content">
+                <div class="projects-group">
+                    <div class="projects-group-title">My Projects</div>
+                    <div class="projects-list">
+                        {loading ? (
+                            <div class="projects-item" style="opacity: 0.7; cursor: default;">
+                                <span class="projects-item-label">Loading...</span>
+                            </div>
+                        ) : filtered.length === 0 ? (
+                            <div class="projects-item" style="opacity: 0.7; cursor: default;">
+                                <span class="projects-item-label">No projects yet</span>
+                            </div>
+                        ) : (
+                            filtered.map((project) => (
+                                <button
+                                    key={project.id}
+                                    type="button"
+                                    class={[
+                                        "projects-item",
+                                        project.id === activeProjectId ? "is-active" : "",
+                                    ].join(" ")}
+                                    onClick={() => onProjectSelect?.(project.id)}
+                                >
+                                    <img src={projectIcon} alt="" class="projects-item-icon" />
+                                    <span class="projects-item-label">{project.name}</span>
+                                </button>
+                            ))
+                        )}
+                    </div>
+                </div>
+            </div>
+
+            {/* New Project button at bottom */}
+            <div class="projects-sidebar-footer">
+                <button
+                    type="button"
+                    class="projects-new-btn"
+                    onClick={() => setShowNewProjectModal(true)}
+                >
+                    <span class="projects-new-plus">＋</span>
+                    New Project
                 </button>
             </div>
 
@@ -111,6 +166,7 @@ function ProjectsList() {
                     onClose={() => setShowNewProjectModal(false)}
                     onProjectCreated={async () => {
                         // Refresh projects list after creating a new one
+                        if (USE_DUMMY_DATA) return;
                         const user = auth.currentUser;
                         if (user) {
                             const userProjects = await listProjects(user.uid);
@@ -119,7 +175,7 @@ function ProjectsList() {
                     }}
                 />
             )}
-        </aside>
+        </>
     );
 }
