+++ frontend/src/pages/Home.tsx
@@ -46,7 +46,6 @@ function Home () {
           </button>
         </section>
       </main>
-      <ProjectsList />
       <SESDCFooter />
     </div>
   );
